Add Admin controller

+++ routes/web.php
@@ -2,12 +2,17 @@
 
 
 // meals controller
-Route::get('/meals','MealController@index')->middleware('auth');
-Route::get('/meals/store','MealController@store')->middleware('auth');
-Route::get('/meals/create','MealController@create')->middleware('auth');
-Route::get('/meals/{id}/edit','MealController@edit')->middleware('auth');
-Route::patch('/meals/{id}','MealController@update')->middleware('auth');
-Route::delete('/products/{id}','ProductController@destroy')->middleware('auth');
+Route::get('/meals','MealController@index') ;
+Route::get('/meals/{id}', 'MealController@show');
+
+
+Route::get('admin/meals/{id}/edit'   ,'Admin\MealController@edit')->middleware('auth');
+Route::patch('admin/meals/{id}'   ,'Admin\MealController@update')->middleware('auth');
+Route::delete('admin/products/{id}'   ,'ProductController@destroy')->middleware('auth');
+Route::get('admin/meals/store'   ,'Admin\MealController@store')->middleware('auth');
+Route::get('admin/meals/create'   ,'Admin\MealController@create')->middleware('auth');
+
+
 
 // category controller
 Route::get('/categories/create','CategoryController@create');
